service/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic-settings==2.*
langchain-core==0.3.*
langchain-google-vertexai==2.*
langchain-openai==0.3.*
# Version MUST match terraform/modules/chromadb/main.tf image tag
chromadb==1.5.0
google-auth==2.*
google-cloud-aiplatform==1.*
google-cloud-bigquery==3.*
